+++ apollo-router-core/src/plugin_utils/mod.rs
@@ -6,6 +6,7 @@ pub use service::{
 };
 pub use structures::{
     execution_request::ExecutionRequest, execution_response::ExecutionResponse,
-    router_request::RouterRequest, router_response::RouterResponse,
-    subgraph_request::SubgraphRequest, subgraph_response::SubgraphResponse,
+    queryplanner_response::QueryPlannerResponse, router_request::RouterRequest,
+    router_response::RouterResponse, subgraph_request::SubgraphRequest,
+    subgraph_response::SubgraphResponse,
 };