+++ apollo-router/src/plugins/mod.rs
@@ -1,2 +1,3 @@
 pub mod override_url;
 mod reporting;
+pub mod rhai;